adiciona a tabela movie ao streaming

+++ src/main/resources/db/migration/V5__create_table_movie_streaming.sql
@@ -1,8 +1,8 @@
 CREATE TABLE IF NOT EXISTS movie_streaming (
-                                               movie_id INTEGER NOT NULL,
-                                               streaming_id INTEGER NOT NULL,
+    movie_id BIGINT NOT NULL,
+    streaming_id BIGINT NOT NULL,
     -- Definindo a Chave Primária Composta (Boa prática para tabelas de junção)
-                                               PRIMARY KEY (movie_id, streaming_id),
+    PRIMARY KEY (movie_id, streaming_id),
     -- Referência correta para a tabela movie
     CONSTRAINT fk_movie_streaming_movie FOREIGN KEY (movie_id) REFERENCES movie(id),
     -- Referência CORRIGIDA para a tabela streaming
